Use BigDecimal as in Transaction to save it properly in database

+++ src/main/java/com/softjourn/coin/server/service/CoinService.java
@@ -270,7 +270,7 @@ public class CoinService {
     }
 
     @SaveTransaction
-    public Transaction deposit(CashDTO cashDTO, String destinationName, String comment, BigInteger amount) {
+    public Transaction deposit(CashDTO cashDTO, String destinationName, String comment, BigDecimal amount) {
         try {
             ErisAccount account = getErisAccount(destinationName);
